refactor: Move `*Frame`
