+++ narwhals/_plan/compliant/typing.py
@@ -7,15 +7,17 @@ from narwhals._typing_compat import TypeVar
 if TYPE_CHECKING:
     from typing_extensions import TypeAlias
 
+    from narwhals._plan.compliant.dataframe import (
+        CompliantBaseFrame,
+        CompliantDataFrame,
+        EagerDataFrame,
+    )
     from narwhals._plan.compliant.group_by import GroupByResolver
     from narwhals._plan.compliant.namespace import CompliantNamespace
     from narwhals._plan.compliant.series import CompliantSeries
     from narwhals._plan.protocols import (
-        CompliantBaseFrame,
-        CompliantDataFrame,
         CompliantExpr,
         CompliantScalar,
-        EagerDataFrame,
         EagerExpr,
         EagerScalar,
         ExprDispatch,
